requirements.txt
# Core dependencies
streamlit==1.28.1
streamlit-chat==0.1.1
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Data processing
numpy==1.24.3
tiktoken==0.5.1

# HTTP client for OpenRouter
requests==2.31.0